Перенес состояние SignIn в HOC
