+++ src/components/movie-card-full-comment/movie-card-full-comment.test.js
@@ -11,7 +11,7 @@ const comment = {
   rating: 8.9,
   text: `Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director's funniest and most exquisitely designed movies in years.`,
   date: `2019-05-08T14:13:56.569Z`
-}
+};
 
 it(`<MovieCardFullComment /> should match snapshot`, () => {
   const tree = renderer
